(fixed osc warnings)
svn-id: https://kwave.svn.sourceforge.net/svnroot/kwave/trunk@2207 bbf32e5a-961d-0410-9e10-91f32741c31e

+++ plugins/codec_wav/WavDecoder.cpp
@@ -274,7 +274,7 @@ bool WavDecoder::open(QWidget *widget, QIODevice &src)
     src.seek(fmt_offset);
 
     // get the encoded block of data from the mime source
-    CHECK(src.size() > sizeof(wav_header_t)+8);
+    CHECK((int)src.size() > (int)sizeof(wav_header_t)+8);
 
     // get the header
     src.read((char *)&header, sizeof(wav_fmt_header_t));
